fix o índice é um float e não um int

+++ web/src/novaLente/novaLente.js
@@ -214,7 +214,7 @@ function cadastrar() {
 		venda : parseFloat(venda), 
 		esferico, 
 		cilindrico, 
-		indice : parseInt(indice), 
+		indice : parseFloat(indice), 
 		diametro : parseInt(diametro) || 0
 	};
